[feature] add new props "enableParseStringFunction"
Hermes will convert the function such as "formatter" to bytecode, and bytecode is not executable in the WebView for Echarts. But if we use a string value for the function, and convert the string value back to a normal function via "new Function" method in the WebView environment. Then Hermes engine will not do any bytecode conversion as the function is created in the WebView environment dynamically. Finally, the function in the option of Echarts will work as expected.